fixes #74: Log warnings when using plugin-provided classes in clustered caches/tasks (#75)
When an instance of a class that is loaded by a PluginClassLoader is used in a Cache or Task, reloading the plugin (and thus replacing the PluginClassLoader) can cause ClassCastExceptions.

This commit detects this usage, and logs warnings. A throttle has been applied to the amount of duplicate warnings logged, to prevent the logs from being flooded with the same message.