fixes #86: Reduce overhead of loading class definitions
This adds a simple cache that stores class definitions loaded by name. As these definitions are looked up for pretty much every serialization action in the cluster, these add significant overhead.

Classes could be loaded/unloaded dynamically. To account for that, the amount of time that class definitions survive in a cache should be limited.

This commit introduces a new property:
- `hazelcast.cache.class.duration` - The maximum duration that a loaded class definition is cached.