improves #86: switch from Caffeine to Openfire cache impl
The previous commit used a Caffeine-based cache implementation, out of fear that using Openfire's Cache implementations would lead to issues (in a clustered environment, the clustered Cache implementation makes use of the functionality that's modified in this change).

By strictly using the local Cache implementation of Openfire, such a conflict (if there ever was one) will not occur.