improves #86: unload classes on plugin changes
Instead of having a very short duration for cached class definitions, this change falls back to the default cache duration, while purging the cache when detecting a plugin change.

Plugin changes are expected to be so infrequent, that a more complex solution (unloading from cache only the classes that are provided by that plugin) will probably introduce more risk than benefit.