OF-2792: Improve cache usage reporting
Clustered Caches should report their actual size, TTL and units used, instead of defaulting to the properties that define non-clustered caches.

See https://igniterealtime.atlassian.net/browse/OF-2792 for a detailed description of the problems addressed by this commit.

This commit assumes that the corresponding Openfire fix for OF-2792 are merged into and released as Openfire 4.8.1.