fixes #103: Prevent race condition on cluster start
A race condition can occur between the cluster start and cluster event listeners being registered with the cluster, causing the listeners to miss events that are sent before the listener registers.

This commit is a first clunky attempt at fixing this problem. It registers the listeners prior to the cluster being started. It then blocks processing of event handling until the cluster is initialized. However, some of this blocking prevents the cluster from initializing, which is why some selected events are mostly ignored.

It's all rather clunky, but it seems to work. I'm committing this as a first working solution, to be improved on further.