fixes #103: Queue cluster events rather than block execution
This builds on the fix to resolve a cluster event listener race, as introduced by the previous commit.

In the previous commit, event handling is blocked until the cluster is fully initialized. A chicken/egg problem (where blockage of certain events prevented the cluster to initialize) caused a haphazardous implementation in which some selected events were _not_ blocked.

In this commit, events are queued rather than blocked. This allows the cluster to initialize, without 'special treatment' of certain events.